Move spring boot related keys in spring boot project
