+++ camel-spring-boot-upgrade-recipes/pom.xml
@@ -22,6 +22,7 @@
     <parent>
         <groupId>org.apache.camel.upgrade</groupId>
         <artifactId>camel-parent-upgrade-recipes</artifactId>
+        <relativePath>../pom.xml</relativePath>
         <version>4.11.0-SNAPSHOT</version>
     </parent>
 
@@ -41,6 +42,22 @@
             <groupId>org.apache.camel.upgrade</groupId>
             <artifactId>camel-upgrade-recipes</artifactId>
             <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openrewrite</groupId>
+            <artifactId>rewrite-yaml</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.openrewrite</groupId>
+            <artifactId>rewrite-properties</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.openrewrite</groupId>
+            <artifactId>rewrite-test</artifactId>
+            <scope>test</scope>
         </dependency>
     </dependencies>
