+++ camel-spring-boot-upgrade-recipes/src/main/resources/META-INF/rewrite/latest.yaml
@@ -30,7 +30,7 @@ recipeList:
   - org.apache.camel.upgrade.camel40.CamelMigrationRecipe
   - org.apache.camel.upgrade.UpgradeToJava17
   - org.apache.camel.upgrade.JavaVersion17
-  - org.apache.camel.upgrade.UpdatePropertiesAndYamlKeys
+  - org.apache.camel.upgrade.UpdateCamelSpringBootPropertiesAndYamlKeys
   - org.openrewrite.maven.UpgradeDependencyVersion:
       groupId: '*camel*'
       artifactId: 'camel-spring-boot-bom'
