--- camel-upgrade-recipes/src/main/resources/META-INF/rewrite/properties.yaml
@@ -1,100 +1,0 @@
-type: specs.openrewrite.org/v1beta/recipe
-name: org.apache.camel.upgrade.UpdatePropertiesAndYamlKeys
-displayName: Update properties and yaml configurations file
-description: Update properties and yaml configurations file.
-recipeList:
-  - org.apache.camel.upgrade.customRecipes.ChangePropertyKeyWithCaseChange:
-      oldPropertyKey: camel.springboot(.*)
-      newPrefix: camel.main
-      exclusions:
-        - camel.springboot.main-run-controller
-        - camel.springboot.include-non-singletons
-        - camel.springboot.warn-on-early-shutdown
-  - org.openrewrite.yaml.ChangeKey:
-      oldKeyPath: $.camel.springboot
-      newKey: main
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerInitialDelay
-      newPropertyKey: camel.routecontroller.initialDelay
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.backlogTracingTemplates
-      newPropertyKey: camel.trace.backlogTracingTemplates
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerThreadPoolSize
-      newPropertyKey: camel.routecontroller.threadPoolSize
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackOffMultiplier
-      newPropertyKey: camel.routecontroller.backOffMultiplier
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.backlogTracing
-      newPropertyKey: camel.trace.enabled
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerExcludeRoutes
-      newPropertyKey: camel.routecontroller.excludeRoutes
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackoffMultiplier
-      newPropertyKey: camel.routecontroller.backOffMultiplier
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.backlogTracingStandby
-      newPropertyKey: camel.trace.backlogTracingStandby
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackOffMaxAttempts
-      newPropertyKey: camel.routecontroller.backOffMaxAttempts
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.springboot.routeControllerSuperviseEnabled
-      newPropertyKey: camel.routecontroller.enabled
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerSuperviseEnabled
-      newPropertyKey: camel.routecontroller.enabled
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackOffMaxDelay
-      newPropertyKey: camel.routecontroller.backOffMaxDelay
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackOffDelay
-      newPropertyKey: camel.routecontroller.backOffDelay
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackoffMaxAttempts
-      newPropertyKey: camel.routecontroller.backOffMaxAttempts
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackoffDelay
-      newPropertyKey: camel.routecontroller.backOffDelay
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackOffMaxElapsedTime
-      newPropertyKey: camel.routecontroller.backOffMaxElapsedTime
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackoffMaxDelay
-      newPropertyKey: camel.routecontroller.backOffMaxDelay
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerIncludeRoutes
-      newPropertyKey: camel.routecontroller.includeRoutes
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.routeControllerBackoffMaxElapsedTime
-      newPropertyKey: camel.routecontroller.backOffMaxElapsedTime
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.backlogTracingTemplates
-      newPropertyKey: camel.trace.backlogTracingTemplates
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.backlogTracingStandby
-      newPropertyKey: camel.trace.backlogTracingStandby
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.backlogTracing
-      newPropertyKey: camel.trace.enabled
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.include-non-singletons
-      newPropertyKey: camel.springboot.include-non-singletons
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.includeNonSingletons
-      newPropertyKey: camel.springboot.include-non-singletons
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.mainRunController
-      newPropertyKey: camel.springboot.main-run-controller
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.main-run-controller
-      newPropertyKey: camel.springboot.main-run-controller
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.warnOnEarlyShutdown
-      newPropertyKey: camel.springboot.warn-on-early-shutdown
-  - org.apache.camel.upgrade.customRecipes.PropertiesAndYamlKeyUpdate:
-      oldPropertyKey: camel.main.warn-on-early-shutdown
-      newPropertyKey: camel.springboot.warn-on-early-shutdown
-
